+++ tests/Infrastructure/Elastica/Transformer/SheetElasticTransformerTest.php
@@ -26,6 +26,7 @@ use Proximum\Vimeet\Domain\Model\Spot;
 use Proximum\Vimeet\Domain\Model\Type;
 use Proximum\Vimeet\Domain\Model\User;
 use Proximum\Vimeet\Domain\Order\Balance;
+use Proximum\Vimeet\Domain\Order\SheetOrderStatus;
 use Proximum\Vimeet\Domain\Repository\CartRowRepositoryInterface;
 use Proximum\Vimeet\Domain\Repository\HappeningParticipationRepositoryInterface;
 use Proximum\Vimeet\Domain\Repository\Invoice\InvoiceRepositoryInterface;
@@ -290,7 +291,6 @@ class SheetElasticTransformerTest extends TestCase
         ;
 
         $balance = $this->prophesize(Balance::class);
-        $balance->getNotCancelledOrderVatViews($sheet->reveal())->shouldBeCalled();
         $balance->getRemainingToPay($sheet->reveal())->shouldBeCalled();
 
         $meetingRepository = $this->prophesize(MeetingRepositoryInterface::class);
@@ -314,6 +314,9 @@ class SheetElasticTransformerTest extends TestCase
             ->willReturn($sheetTemplateWithTaggedDataEn)
         ;
 
+        $sheetOrderStatus = $this->prophesize(SheetOrderStatus::class);
+        $sheetOrderStatus->getStatus($sheet->reveal())->shouldBeCalled()->willReturn(Sheet\Constant::NO_ORDER);
+
         $transformer = new SheetElasticTransformer(
             $sheetInfoGuesser->reveal(),
             $participantInfoGuesser->reveal(),
@@ -325,6 +328,7 @@ class SheetElasticTransformerTest extends TestCase
             $balance->reveal(),
             $meetingRepository->reveal(),
             $invoiceRepository->reveal(),
+            $sheetOrderStatus->reveal(),
             $campaignRepository->reveal()
         );